Finalized exception preview in gui
