+++ gui/src/App/Components/Header/Header.scss
@@ -1,4 +1,4 @@
-@import '../Misc/colors';
+@import '../../Misc/colors';
 
 $height: 3rem;
